requirements.txt
# TrendStack Trading Environment - Requirements
# Core trading libraries: backtrader, pandas, numpy, scipy, yfinance
# Technical analysis: ta (Technical Analysis library)
# Visualization: matplotlib, plotly, seaborn, quantstats
# Data processing: pydantic, pyyaml, requests
# Utilities: loguru, schedule, prefect
# Communication: python-telegram-bot
# Trading platforms: MetaTrader5
# Note: empyrical was skipped due to Python 3.13 compatibility issues
# Note: vectorbt-lite was not available

aiosqlite==0.21.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
apprise==1.9.3
APScheduler==3.6.3
asgi-lifespan==2.1.0
asyncpg==0.30.0
attrs==25.3.0
backtrader==1.9.78.123
beautifulsoup4==4.13.4
cachetools>=4.2.2,<7.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
coolname==2.2.0
cryptography==45.0.5
curl_cffi==0.12.0
cycler==0.12.1
dateparser==1.2.2
docker==7.1.0
exceptiongroup==1.3.0
fastapi==0.116.1
fonttools==4.59.0
frozendict==2.4.6
fsspec==2025.7.0
graphviz==0.21
greenlet==3.2.3
griffe==1.8.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
hyperframe==6.1.0
idna==3.10
importlib-metadata==8.7.0
jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
loguru==0.7.3
Mako==1.3.10
markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
MetaTrader5==5.0.5120
multitasking==0.0.12
narwhals==1.48.0
numpy==2.3.1
oauthlib==3.3.1
opentelemetry-api==1.35.0
orjson==3.11.0
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
prefect==3.4.10
prometheus-client==0.22.1
protobuf==6.31.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
python-socks==2.7.1
python-telegram-bot>=20.0
pytz==2025.2
pywin32==311
PyYAML==6.0.2
QuantStats==0.0.68
readchar==4.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.0.0
rpds-py==0.26.0
ruamel-yaml==0.18.14
ruamel.yaml.clib==0.2.12
schedule==1.2.2
scipy==1.16.0
seaborn==0.13.2
semver==3.0.4
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.41
starlette==0.47.2
ta==0.11.0
tabulate==0.9.0
text-unidecode==1.3
toml==0.10.2
tornado==6.1
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uv==0.8.2
uvicorn==0.35.0
websockets==15.0.1
whenever==0.8.7
win32_setctime==1.2.0
yfinance==0.2.65
zipp==3.23.0
